requirements/base.txt
asgiref==3.5.0
astroid==2.9.3
cfgv==3.3.1
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.2
identify==2.4.8
lazy-object-proxy==1.7.1
mccabe==0.6.1
mysqlclient==2.1.0
nodeenv==1.6.0
platformdirs==2.4.1
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.13.0
wrapt==1.13.3
